Move DEFAULT_EDITOR_ASSOCIATION
